third_party/WebKit/LayoutTests/http/tests/devtools/sources/outline-javascript-6-expected.txt

Verify javascript outline


Running: testClassAsNumberLiteralProperty
Text:
    var foo = { 42: class { }};
Outline:

Running: testClassAsStringLiteralProperty
Text:
    var foo = { "foo": class { }};
Outline:

Running: testClassAsIdentifierProperty
Text:
    var foo = { foo: class { }};
Outline:
    class foo:0:12
